refactor: better architecture and film storage
