+++ cmd/botapp/main.go
@@ -16,13 +16,13 @@ import (
 
 	"github.com/HalvaPovidlo/halvabot-go/cmd/config"
 	v1 "github.com/HalvaPovidlo/halvabot-go/internal/api/v1"
-	"github.com/HalvaPovidlo/halvabot-go/internal/api/v1/login"
-	musicrest "github.com/HalvaPovidlo/halvabot-go/internal/api/v1/music"
-	capi "github.com/HalvaPovidlo/halvabot-go/internal/chess/api/discord"
+	v1chess "github.com/HalvaPovidlo/halvabot-go/internal/api/v1/chess"
+	v1login "github.com/HalvaPovidlo/halvabot-go/internal/api/v1/login"
+	v1music "github.com/HalvaPovidlo/halvabot-go/internal/api/v1/music"
 	"github.com/HalvaPovidlo/halvabot-go/internal/chess/lichess"
+	"github.com/HalvaPovidlo/halvabot-go/internal/login"
 	"github.com/HalvaPovidlo/halvabot-go/internal/music"
 	"github.com/HalvaPovidlo/halvabot-go/internal/music/audio"
-	dapi "github.com/HalvaPovidlo/halvabot-go/internal/music/discord"
 	"github.com/HalvaPovidlo/halvabot-go/internal/music/player"
 	ytsearch "github.com/HalvaPovidlo/halvabot-go/internal/music/search/youtube"
 	"github.com/HalvaPovidlo/halvabot-go/internal/music/storage/firestore"
@@ -84,24 +84,25 @@ func main() {
 	}
 
 	// Music stage
-	voiceClient := audio.NewVoiceClient(session)
-	rawAudioPlayer := audio.NewPlayer(loadMaster, &cfg.Discord.Voice.EncodeOptions)
-	musicPlayer := player.NewMusicService(ctx, fireService, ytClient, voiceClient, rawAudioPlayer)
+	musicService := v1music.NewMusicService(
+		ctx,
+		fireService,
+		ytClient,
+		player.NewPlayer(ctx, audio.NewVoiceClient(session), audio.NewPlayer(loadMaster, &cfg.Discord.Voice.EncodeOptions)),
+	)
 
 	// Chess
 	lichessClient := lichess.NewClient()
 
 	// Discord commands
-	musicCog := dapi.NewCog(musicPlayer, cfg.Discord.Prefix, cfg.Discord.API)
-	musicCog.RegisterCommands(ctx, session, cfg.General.Debug, logger)
-	chessCog := capi.NewCog(cfg.Discord.Prefix, lichessClient)
-	chessCog.RegisterCommands(session, cfg.General.Debug, logger)
+	v1music.NewDiscordMusicHandler(musicService, cfg.Discord.Prefix, cfg.Discord.API).RegisterCommands(ctx, session, cfg.General.Debug, logger)
+	v1chess.NewDiscordChessHandler(cfg.Discord.Prefix, lichessClient).RegisterCommands(session, cfg.General.Debug, logger)
 
 	// Auth stage
-	loginService := login.NewLoginService(login.NewAccountStorage(fireClient), jwt.NewJWTokenizer(cfg.Secret))
+	loginService := v1login.NewLoginHandler(login.NewAccountStorage(fireClient), jwt.NewJWTokenizer(cfg.Secret))
 
 	// Http routers
-	server := v1.NewServer(musicrest.NewMusicHandler(musicPlayer, logger), loginService)
+	server := v1.NewServer(v1music.NewMusicHandler(musicService, logger), loginService)
 	server.Run(cfg.Host.IP, cfg.Host.Bot, config.SwaggerPath, cfg.General.Debug)
 
 	sc := make(chan os.Signal, 1)
