+++ cmd/botmock/main.go
@@ -9,9 +9,9 @@ import (
 
 	"github.com/HalvaPovidlo/halvabot-go/cmd/config"
 	v1 "github.com/HalvaPovidlo/halvabot-go/internal/api/v1"
-	"github.com/HalvaPovidlo/halvabot-go/internal/api/v1/login"
+	v1login "github.com/HalvaPovidlo/halvabot-go/internal/api/v1/login"
 	"github.com/HalvaPovidlo/halvabot-go/internal/api/v1/music"
-	"github.com/HalvaPovidlo/halvabot-go/internal/music/player"
+	"github.com/HalvaPovidlo/halvabot-go/internal/login"
 	"github.com/HalvaPovidlo/halvabot-go/pkg/http/jwt"
 	"github.com/HalvaPovidlo/halvabot-go/pkg/log"
 )
@@ -23,9 +23,8 @@ func main() {
 	}
 	logger := log.NewLogger(cfg.General.Debug)
 
-	mock := &player.MockPlayer{}
-	loginService := login.NewLoginService(login.NewMockStorage(), jwt.NewJWTokenizer("mock_secret"))
-	musicService := music.NewMusicHandler(mock, logger)
+	loginService := v1login.NewLoginHandler(login.NewMockStorage(), jwt.NewJWTokenizer("mock_secret"))
+	musicService := music.NewMusicHandler(&music.MockPlayer{}, logger)
 	// Http routers
 	server := v1.NewServer(musicService, loginService)
 	server.Run(cfg.Host.IP, cfg.Host.Mock, config.SwaggerPath, cfg.General.Debug)
