+++ internal/api/v1/music/discord_messages.go
@@ -1,4 +1,4 @@
-package discord
+package music
 
 import (
 	"context"
@@ -32,8 +32,8 @@ const (
 	infoLevel
 )
 
-func (s *Service) sendComplexMessage(ctx context.Context, session *dg.Session, channelID string, msg *dg.MessageSend, level int) {
-	if s.toDelete(channelID, level) {
+func (h *DiscordHandler) sendComplexMessage(ctx context.Context, session *dg.Session, channelID string, msg *dg.MessageSend, level int) {
+	if h.toDelete(channelID, level) {
 		return
 	}
 	go func() {
@@ -47,48 +47,48 @@ func (s *Service) sendComplexMessage(ctx context.Context, session *dg.Session, c
 	}()
 }
 
-func (s *Service) sendSearchingMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate) {
-	s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageSearching), statusLevel)
+func (h *DiscordHandler) sendSearchingMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate) {
+	h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageSearching), statusLevel)
 }
 
-func (s *Service) sendSkipMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate) {
-	s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageSkip), statusLevel)
+func (h *DiscordHandler) sendSkipMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate) {
+	h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageSkip), statusLevel)
 }
 
-func (s *Service) sendFoundMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, artist, title string, playbacks int) {
+func (h *DiscordHandler) sendFoundMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, artist, title string, playbacks int) {
 	msg := fmt.Sprintf("%s `%s - %s` %s", messageFound, artist, title, intToEmoji(playbacks))
-	s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(msg), statusLevel)
+	h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(msg), statusLevel)
 }
 
-func (s *Service) sendNotFoundMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate) {
-	s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageNotFound), statusLevel)
+func (h *DiscordHandler) sendNotFoundMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate) {
+	h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageNotFound), statusLevel)
 }
 
-func (s *Service) sendAgeRestrictionMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate) {
-	s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageAgeRestriction), statusLevel)
+func (h *DiscordHandler) sendAgeRestrictionMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate) {
+	h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageAgeRestriction), statusLevel)
 }
 
-func (s *Service) sendLoopMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, enabled bool) {
+func (h *DiscordHandler) sendLoopMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, enabled bool) {
 	if enabled {
-		s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageLoopEnabled), statusLevel)
+		h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageLoopEnabled), statusLevel)
 	} else {
-		s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageLoopDisabled), statusLevel)
+		h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageLoopDisabled), statusLevel)
 	}
 }
 
-func (s *Service) sendRadioMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, enabled bool) {
+func (h *DiscordHandler) sendRadioMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, enabled bool) {
 	if enabled {
-		s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageRadioEnabled), statusLevel)
+		h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageRadioEnabled), statusLevel)
 	} else {
-		s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageRadioDisabled), statusLevel)
+		h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageRadioDisabled), statusLevel)
 	}
 }
 
-func (s *Service) sendNotInVoiceWarning(ctx context.Context, ds *dg.Session, m *dg.MessageCreate) {
-	s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageNotVoiceChannel), statusLevel)
+func (h *DiscordHandler) sendNotInVoiceWarning(ctx context.Context, ds *dg.Session, m *dg.MessageCreate) {
+	h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(messageNotVoiceChannel), statusLevel)
 }
 
-func (s *Service) sendNowPlayingMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, song *item.Song, pos float64) {
+func (h *DiscordHandler) sendNowPlayingMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, song *item.Song, pos float64) {
 	msg := &dg.MessageSend{
 		Embeds: []*dg.MessageEmbed{
 			{
@@ -123,35 +123,35 @@ func (s *Service) sendNowPlayingMessage(ctx context.Context, ds *dg.Session, m *
 			},
 		},
 	}
-	s.sendComplexMessage(ctx, ds, m.ChannelID, msg, infoLevel)
+	h.sendComplexMessage(ctx, ds, m.ChannelID, msg, infoLevel)
 }
 
-func (s *Service) sendRandomMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, songs []*item.Song) {
+func (h *DiscordHandler) sendRandomMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, songs []*item.Song) {
 	msg := ""
 	for _, song := range songs {
 		if song.ArtistName != "" {
-			msg += fmt.Sprintf("`%s%s - %s`\n", s.prefix+play, song.ArtistName, song.Title)
+			msg += fmt.Sprintf("`%s%s - %s`\n", h.prefix+play, song.ArtistName, song.Title)
 		} else {
-			msg += fmt.Sprintf("`%s%s`\n", s.prefix+play, song.Title)
+			msg += fmt.Sprintf("`%s%s`\n", h.prefix+play, song.Title)
 		}
 	}
-	s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(msg), infoLevel)
+	h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(msg), infoLevel)
 }
 
-func (s *Service) sendInternalErrorMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, level int) {
-	s.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(discord.MessageInternalError), level)
+func (h *DiscordHandler) sendInternalErrorMessage(ctx context.Context, ds *dg.Session, m *dg.MessageCreate, level int) {
+	h.sendComplexMessage(ctx, ds, m.ChannelID, strmsg(discord.MessageInternalError), level)
 }
 
-// func (s *Service) sendStringMessage(ds *dg.Session, m *dg.MessageCreate, msg string, level int) {
+// func (s *DiscordHandler) sendStringMessage(ds *dg.Session, m *dg.MessageCreate, msg string, level int) {
 //	s.sendComplexMessage(ds, m.ChannelID, strmsg(msg), level)
 // }
 
-func (s *Service) toDelete(channelID string, level int) bool {
-	s.channelsMx.RLock()
-	name := s.allChannels[channelID]
-	_, status := s.statusChannels[name]
-	_, open := s.openChannels[name]
-	s.channelsMx.RUnlock()
+func (h *DiscordHandler) toDelete(channelID string, level int) bool {
+	h.channelsMx.RLock()
+	name := h.allChannels[channelID]
+	_, status := h.statusChannels[name]
+	_, open := h.openChannels[name]
+	h.channelsMx.RUnlock()
 	if level <= infoLevel && !(open || status) {
 		return true
 	}
@@ -161,10 +161,10 @@ func (s *Service) toDelete(channelID string, level int) bool {
 	return false
 }
 
-func (s *Service) loadChannelsID(ds *dg.Session, guildID string) {
-	s.channelsMx.Lock()
-	defer s.channelsMx.Unlock()
-	if len(s.allChannels) != 0 {
+func (h *DiscordHandler) loadChannelsID(ds *dg.Session, guildID string) {
+	h.channelsMx.Lock()
+	defer h.channelsMx.Unlock()
+	if len(h.allChannels) != 0 {
 		return
 	}
 
@@ -173,7 +173,7 @@ func (s *Service) loadChannelsID(ds *dg.Session, guildID string) {
 		return
 	}
 	for _, v := range channels {
-		s.allChannels[v.ID] = v.Name
+		h.allChannels[v.ID] = v.Name
 	}
 }
