+++ internal/api/v1/music/mock.go
@@ -1,4 +1,4 @@
-package player
+package music
 
 import (
 	"context"
