--- internal/film/api/rest/handlers.go
@@ -1,1 +1,0 @@
-package rest
