+++ internal/login/mock.go
@@ -2,6 +2,7 @@ package login
 
 import (
 	"context"
+	login2 "github.com/HalvaPovidlo/halvabot-go/internal/api/v1/login"
 )
 
 type mock struct{}
@@ -11,8 +12,8 @@ func NewMockStorage() *mock {
 }
 
 func (a *mock) GetAccount(ctx context.Context, login string) (*AccountInfo, error) {
-	return &AccountInfo{
-		Password: "password",
-		UserID:   "user_id",
+	return &login.AccountInfo{
+		Password:      "password",
+		login2.UserID: "user_id",
 	}, nil
 }
